requirements.txt
certifi==2020.12.5
chardet==4.0.0
google-trans-new==1.1.9
idna==2.10
mysql-connector==2.2.9
mysql-connector-python==8.0.24
protobuf==3.15.8
pyTelegramBotAPI==3.7.7
python-dotenv==0.17.0
requests==2.25.1
six==1.15.0
urllib3==1.26.4
